fix: refactor selectBusinessUnitsUI
