--- src/hooks/useSelectBusinessUnits.ts
@@ -1,34 +1,0 @@
-import { useEffect } from "react";
-import { useNavigate, useLocation } from "react-router-dom";
-import { useContext } from "react";
-import { validateBusinessUnities } from "@pages/selectBusinessUnits/utils";
-import { AuthAndData } from "@context/authAndDataProvider";
-
-const useSelectBusinessUnits = () => {
-  const navigate = useNavigate();
-  const location = useLocation();
-  const { appData, setBusinessUnitsToTheStaff } = useContext(AuthAndData);
-
-  useEffect(() => {
-    if (appData.portal.publicCode) {
-      validateBusinessUnities(
-        appData.portal.publicCode,
-        appData.user.userAccount
-      ).then((data) => {
-        setBusinessUnitsToTheStaff(data);
-      });
-    }
-  }, [appData.portal.publicCode, setBusinessUnitsToTheStaff]);
-
-  useEffect(() => {
-    if (
-      location.pathname === "/selectBusinessUnit" ||
-      location.pathname === "/selectBusinessUnit/" ||
-      location.pathname === "/"
-    ) {
-      navigate(`/selectBusinessUnit/checking-credentials/`);
-    }
-  }, [location, navigate]);
-};
-
-export { useSelectBusinessUnits };
